fix(angular): modify angularjs tests
