+++ e2e/angular/src/angular-package.test.ts
@@ -144,8 +144,8 @@ forEachCli('angular', (cli) => {
         [`@proj/${childLib2}`]: '0.0.1',
       });
       expect(jsonFile.peerDependencies).toEqual({
-        '@angular/common': '^10.0.0',
-        '@angular/core': '^10.0.0',
+        '@angular/common': '^10.0.2',
+        '@angular/core': '^10.0.2',
       });
     });
   });
