+++ packages/angular/migrations.json
@@ -31,5 +31,16 @@
       "description": "Upgrades Angular and Angular CLI to 10.0.0",
       "factory": "./src/migrations/update-10-0-0/update-10-0-0"
     }
+  },
+  "packageJsonUpdates": {
+    "10.0.0": {
+      "version": "10.0.0-beta.0",
+      "packages": {
+        "rxjs": {
+          "version": "~6.5.5",
+          "alwaysAddToPackageJson": false
+        }
+      }
+    }
   }
 }
