+++ packages/angular/src/utils/versions.ts
@@ -3,5 +3,5 @@ export const angularVersion = '^10.0.0';
 export const angularDevkitVersion = '~0.1000.0';
 export const angularJsVersion = '1.7.9';
 export const ngrxVersion = '9.1.0';
-export const rxjsVersion = '~6.5.4';
+export const rxjsVersion = '~6.5.5';
 export const jestPresetAngularVersion = '8.1.2';
