+++ packages/nest/migrations.json
@@ -65,7 +65,7 @@
           "alwaysAddToPackageJson": false
         },
         "rxjs": {
-          "version": "^6.5.4",
+          "version": "~6.5.5",
           "alwaysAddToPackageJson": true
         }
       }
