+++ packages/nest/src/utils/versions.ts
@@ -3,6 +3,6 @@ export const nxVersion = '*';
 export const nestJsVersion = '^7.0.0';
 export const nestJsSchematicsVersion = '^7.0.0';
 
-export const rxjsVersion = '^6.5.4';
+export const rxjsVersion = '~6.5.5';
 
 export const reflectMetadataVersion = '^0.1.13';
