+++ packages/next/src/schematics/application/lib/add-jest.ts
@@ -6,7 +6,6 @@ import {
 } from '@angular-devkit/schematics';
 import { updateJsonInTree } from '@nrwl/workspace';
 import { NormalizedSchema } from './normalize-options';
-import { join } from '@angular-devkit/core';
 
 export function addJest(options: NormalizedSchema): Rule {
   return options.unitTestRunner === 'jest'
@@ -17,21 +16,11 @@ export function addJest(options: NormalizedSchema): Rule {
           skipSerializers: true,
           setupFile: 'none',
         }),
+
         updateJsonInTree(
-          join(options.appProjectRoot, 'tsconfig.spec.json'),
+          `${options.appProjectRoot}/tsconfig.spec.json`,
           (json) => {
-            json.compilerOptions = {
-              ...json.compilerOptions,
-              jsx: 'react',
-              module: 'commonjs',
-              allowJs: true,
-              esModuleInterop: true,
-              allowSyntheticDefaultImports: true,
-              forceConsistentCasingInFileNames: true,
-              noEmit: true,
-              resolveJsonModule: true,
-              isolatedModules: true,
-            };
+            json.compilerOptions.jsx = 'react';
             return json;
           }
         ),
