+++ packages/next/src/schematics/application/lib/update-jest-config.ts
@@ -1,6 +1,5 @@
-import { chain, noop, Rule } from '@angular-devkit/schematics';
+import { noop, Rule } from '@angular-devkit/schematics';
 import { NormalizedSchema } from './normalize-options';
-import { offsetFromRoot, updateJsonInTree } from '@nrwl/workspace';
 
 export function updateJestConfig(options: NormalizedSchema): Rule {
   return options.unitTestRunner === 'none'
